# docs/swagger.yaml
basePath: /v1
definitions:
  entity.Animal:
    properties:
      animal:
        example: sheep
        type: string
      id:
        type: string
      is_hungry:
        example: true
        type: boolean
      weight:
        example: 56.9
        type: number
    type: object
  models.CreateAnimalRequest:
    properties:
      animal:
        example: sheep
        type: string
      is_hungry:
        example: true
        type: boolean
      weight:
        example: 56.9
        type: number
    type: object
  v1.response:
    properties:
      error:
        example: message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Dostonbek Soliyev, Nodirbek No'monov
  title: Farmish APP
  version: "1.0"
paths:
  /animal/create:
    post:
      consumes:
      - application/json
      description: Api for creating animal
      operationId: animal-create
      parameters:
      - description: Admin credentials for logging in
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAnimalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Animal'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create Animal
      tags:
      - animal
  /animal/update/{id}:
    put:
      consumes:
      - application/json
      description: Api for updating animal
      operationId: animal-update
      parameters:
      - description: Animal ID
        in: path
        name: id
        required: true
        type: string
      - description: Admin credentials for logging in
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAnimalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Animal'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update Animal
      tags:
      - animal
swagger: "2.0"